Add files via upload
this project, I built a model to predict the rating of a restaurant based on the reviews. The model was trained on a dataset of over 20,000 reviews, and it achieved an accuracy of 85%.

The project was divided into the following steps:

Data cleaning: The data was cleaned by removing stop words and performing lemmatization.
Feature extraction: A TF-IDF vectorizer was used to create a numerical representation of the reviews.
Model training: A logistic regression model was trained on the cleaned and extracted data.
Model evaluation: The model was evaluated on a test set, and it achieved an accuracy of 85%.
Cross-validation: The model was cross-validated to evaluate its generalizability. The cross-validation accuracy was 83%.
The results of the project show that the model is able to predict the rating of a restaurant with a reasonable degree of accuracy. However, there is still room for improvement. In the future, I would like to explore other machine-learning models and try to improve the accuracy of the predictions.

Libraries Used:

pandas
NumPy
matplotlib
seaborn
nltk
sklearn
Conclusion:

The project was a success, and the model was able to predict the rating of a restaurant with a reasonable degree of accuracy. However, there is still room for improvement, and I would like to explore other machine-learning models in the future.